✨ Refactor MCPClient to use context managers for connection management, enhancing resource cleanup and error handling. Update demo scripts to reflect new MCPClient usage patterns with improved transport connection methods. Remove legacy MCPClient implementation from fence.tools.mcp.client.
